+++ test/plugins/iamauth.js
@@ -604,4 +604,58 @@ describe('#db IAMAuth Plugin', function() {
       done();
     });
   });
+
+  it('supports changing the IAM API key', function(done) {
+    if (process.env.NOCK_OFF) {
+      this.skip();
+    }
+
+    var iamMocks = nock(TOKEN_SERVER)
+      .post('/identity/token', {
+        'grant_type': 'urn:ibm:params:oauth:grant-type:apikey',
+        'response_type': 'cloud_iam',
+        'apikey': 'bad_key'
+      })
+      .reply(400, {
+        errorCode: 'BXNIM0415E',
+        errorMessage: 'Provided API key could not be found'
+      })
+      .post('/identity/token', {
+        'grant_type': 'urn:ibm:params:oauth:grant-type:apikey',
+        'response_type': 'cloud_iam',
+        'apikey': IAM_API_KEY
+      })
+      .reply(200, MOCK_IAM_TOKEN_RESPONSE);
+
+    var cloudantMocks = nock(SERVER)
+      .get(DBNAME)
+      .reply(401, {error: 'unauthorized', reason: 'Unauthorized'})
+      .post('/_iam_session', {access_token: MOCK_ACCESS_TOKEN})
+      .reply(200, {ok: true}, MOCK_SET_IAM_SESSION_HEADER)
+      .get(DBNAME)
+      .reply(200, {doc_count: 0});
+
+    var cloudantClient = new Client({ plugins: { iamauth: { iamApiKey: 'bad_key' } } });
+    var req = { url: SERVER + DBNAME, method: 'GET' };
+    cloudantClient.request(req, function(err, resp, data) {
+      assert.equal(err, null);
+      assert.equal(resp.request.headers.cookie, null);
+      assert.equal(resp.statusCode, 401);
+      assert.ok(data.indexOf('"error":"unauthorized"') > -1);
+
+      // update IAM API key
+      cloudantClient.getPlugin('iamauth').setIamApiKey(IAM_API_KEY);
+
+      cloudantClient.request(req, function(err, resp, data) {
+        assert.equal(err, null);
+        assert.equal(resp.request.headers.cookie, MOCK_IAM_SESSION);
+        assert.equal(resp.statusCode, 200);
+        assert.ok(data.indexOf('"doc_count":0') > -1);
+
+        iamMocks.done();
+        cloudantMocks.done();
+        done();
+      });
+    });
+  });
 });