+++ CHANGES.md
@@ -1,4 +1,5 @@
 # UNRELEASED
+- [NEW] Added option to set new IAM API key.
 - [FIXED] Allow plugins to be loaded from outside the 'plugins/' directory.
 - [FIXED] Retry bad IAM token requests.